daily_storage/requirements.txt
s3fs
psycopg2-binary
datetime
python-dotenv